\name{rgph2024tn-package}
\alias{rgph2024tn-package}
\docType{package}
\title{Tunisia 2024 General Census of Population and Housing}
\description{Harmonized data from the 2024 RGPH: 24 gouvernorats,
  279 delegations, 2082 secteurs, 46 indicators. Source: INS.}
\examples{
codebook()
del <- load_delegations()
nrow(del)
}
